Rename package in quarkus project
Resolves #153
