+++ platform/quarkus/src/test/java/org/trellisldp/ext/quarkus/AbstractApplicationTests.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.trellisldp.ext.webapp;
+package org.trellisldp.ext.quarkus;
 
 import static io.restassured.RestAssured.get;
 import static io.restassured.RestAssured.given;
